requirements.txt
streamlit 
langchain 
openai 
python-dotenv 
pandas 
requests 
beautifulsoup4
google-generativeai